+++ client/common/communication.go
@@ -5,7 +5,24 @@ import (
 	"net"
     "time"
     "fmt"
+    "strings"
 )
+
+func ReadResponse(reader *bufio.Reader, clientID string) (string, error) {
+	response, err := reader.ReadString('\n')
+	if err != nil {
+		log.Errorf("action: read_response | result: fail | client_id: %v | error: %v", clientID, err)
+		return "", err
+	}
+
+	trimmedResponse := strings.TrimSpace(response)
+
+	if trimmedResponse == "BATCH_ERROR" || trimmedResponse == "ERROR_BATCH" {
+		return trimmedResponse, fmt.Errorf("server returned batch error")
+	}
+
+	return trimmedResponse, nil
+}
 
 func SendMessage(conn net.Conn, message string, clientID string) error {
     msgBytes := []byte(message)
